pyproject.toml
[project]
name = "article_bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Stéphane Burwash",email = "sburwash@live.ca"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-core (>=0.3.51,<0.4.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
